refactored components and added filter component
