+++ src/screens/Analytics/Components/Chart/Chart.scss
