+++ src/screens/Analytics/Components/Filter/Filter.scss
@@ -1,0 +1,3 @@
+.filter-container {
+  margin-left: 6rem;
+}
