+++ src/screens/Analytics/Components/Tooltip/CustomTooltip.scss
@@ -1,0 +1,9 @@
+.tooltip-box {
+  font-size: 12px;
+  border: 1px solid black;
+  background-color: white;
+}
+
+.text-box {
+  margin: 1rem;
+}
