+++ src/screens/Analytics/mockData.js
@@ -1,0 +1,70 @@
+export default {
+  totalTimeWatched: 120,
+  data: [
+    {
+      courseName: 'Artificial Intelligence',
+      watchHistory: [
+        {
+          date: '06/09/2020',
+          duration: 40,
+          videoName: 'Reinforcement Learning',
+        },
+        {
+          date: '06/08/2020',
+          duration: 20,
+          videoName: 'Q-Learning',
+        },
+        {
+          date: '06/07/2020',
+          duration: 50,
+          videoName: 'Minimax',
+        },
+      ],
+    },
+    {
+      courseName: 'Machine Learning',
+      watchHistory: [
+        {
+          date: '06/09/2020',
+          duration: 20,
+          videoName: 'Gradient Descent',
+        },
+        {
+          date: '06/08/2020',
+          duration: 70,
+          videoName: 'SVMs',
+        },
+        {
+          date: '06/07/2020',
+          duration: 35,
+          videoName: 'KNN',
+        },
+        {
+          date: '06/06/2020',
+          duration: 100,
+          videoName: 'More',
+        },
+        {
+          date: '06/05/2020',
+          duration: 20,
+          videoName: 'More 1',
+        },
+        {
+          date: '06/04/2020',
+          duration: 25,
+          videoName: 'More 2',
+        },
+        {
+          date: '06/03/2020',
+          duration: 50,
+          videoName: 'More 3',
+        },
+        {
+          date: '06/02/2020',
+          duration: 35,
+          videoName: 'More 4',
+        },
+      ],
+    },
+  ],
+};
